Lemay task 2/requirements.txt
fastapi
uvicorn
transformers
torch
accelerate
gunicorn
python-dotenv